Foundation: Control Flow #time 15m

+++ control-flow/solution/ifelse_input.py
@@ -3,7 +3,7 @@ AGE_ADULT = 18
 
 age = input('What is your age?: ')
 
-if float(age) >= AGE_ADULT:
+if int(age) >= AGE_ADULT:
     print('Adult')
 else:
     print('Young')
